Security

+++ client/src/gui/popups/updater.c
@@ -87,6 +87,8 @@
 #include <global.h>
 #include <toolkit_string.h>
 #include <curl.h>
+#include <sha1.h>
+#include <path.h>
 
 /** Holds the current cURL request that is being processed. */
 static curl_request_t *request = NULL;
@@ -264,9 +266,9 @@ updater_process_packages (void)
 
             /* Construct the path. */
             char filename[HUGE_BUF];
-            snprintf(VS(filename), "%s/client_patch_%09zu.tar.gz",
+            snprintf(VS(filename), "%s/client_patch_%09" PRIu64 ".tar.gz",
                      dir_path,
-                     download_package_next - 1);
+                     (uint64_t) download_package_next - 1);
 
             path_ensure_directories(filename);
             FILE *fp = fopen(filename, "wb");
